added patch and delete routes for cards

+++ src/mini_blog_api/repositories/card_repository.py
@@ -2,13 +2,14 @@ from typing import Any, Dict, List, Optional
 from datetime import datetime
 from bson.objectid import ObjectId
 
-import structlog
 import json
+import structlog
 from fastapi import HTTPException
 from motor.motor_asyncio import AsyncIOMotorClient, AsyncIOMotorCursor
 from pymongo.errors import ServerSelectionTimeoutError
 
 from ..models.card_model import Card, CardPayloadCreate
+from .author_repository import UserRepository
 from ..services.util import sanitize
 
 log = structlog.get_logger()
@@ -18,18 +19,19 @@ class CardRepository:
         cls.collection = db["cards"]
     
     @classmethod
-    async def find_card(cls, name):
+    async def find_card(cls, card_name: str):
         try:
-            user_dict: Dict[str, Any] = await cls.collection.find_one(dict(name=name))
-            if user_dict:
-                return Card.model_validate(user_dict)
+            card_dict: Dict[str, Any] = await cls.collection.find_one(dict(name=card_name))
+            
+            if card_dict:
+                return Card.model_validate(card_dict)
             
         except ServerSelectionTimeoutError as error:
             log.msg(error)
             raise HTTPException(500, "Failed to connect to MongoDB.")
         
     @classmethod
-    async def find_category_list(cls, query_filter: Dict[str, Any]) -> List[Card]:
+    async def find_card_list(cls, query_filter: Dict[str, Any]) -> List[Card]:
         try:
             docs: List[Card] = []
             db_cursor: AsyncIOMotorCursor = cls.collection.find(query_filter)
@@ -60,15 +62,22 @@ class CardRepository:
             raise HTTPException(500, "Failed to connect to MongoDB.")
         
     @classmethod
-    async def update_card(cls, card_name: str, card_data: str, author: str) -> Optional[Dict]:
+    async def update_card(cls, card_name: str, card_data: Dict[str, Any], author: str) -> Optional[Dict]:
         try:
             card = await cls.find_card(card_name)
-
+            
             if not card:
                 return None
             
-            if card.author != author:
-                raise HTTPException(status_code=403, detail="You are not authorized to edit this card")
+            author_detail = await UserRepository.find_user_by_id(card.author)
+
+            if author_detail.username != author:
+                raise HTTPException(status_code=403, detail="You are unauthorized to edit this card")
+            
+            query = {"name" : card_name}
+            set_data = {"$set": card_data}
+            
+            return await cls.collection.update_one(query, set_data)
             
         except ServerSelectionTimeoutError as error:
             log.msg(error)
@@ -76,9 +85,18 @@ class CardRepository:
         
     
     @classmethod
-    async def delete_card(cls, card_name: str):
+    async def delete_card(cls, card_name: str, author: str):
         try:
-           await cls.collection.delete_one(dict(name=card_name))
+           card = await cls.find_card(card_name)
+           if not card:
+                return None
+           
+           author_detail = await UserRepository.find_user_by_id(card.author)
+           
+           if author_detail.username != author:
+                raise HTTPException(status_code=403, detail="You are unauthorized to delete this card")
+           
+           return await cls.collection.delete_one(dict(name=card_name))
             
         except ServerSelectionTimeoutError as error:
             log.msg(error)
